+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes to babel-plugin-import-postcss
 
+### 2.0.0 (May 24, 2019)
+
+- Updated: Node 8+ compatibility (major)
+- Updated: `cosmiconfig` to 5.2.1 (patch)
+- Updated: `deasync` to 0.1.15 (patch)
+- Updated: `postcss` to 7.0.16 (patch)
+
 ### 1.2.0 (March 27, 2019)
 
 - Updated: `comsmiconfig` to 5.2.0 (minor)
