+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "babel-plugin-import-postcss",
-  "version": "1.2.0",
+  "version": "2.0.0",
   "description": "Import processed CSS files in JS",
   "author": "Jonathan Neal <jonathantneal@hotmail.com>",
   "license": "CC0-1.0",
@@ -19,30 +19,43 @@
     "build": "rollup --config --silent",
     "prepublishOnly": "npm test && npm run build",
     "test": "npm run test:js",
-    "test:js": "eslint src/*.js --cache --ignore-path .gitignore --quiet"
+    "test:js": "eslint src/{*,**/*}.js --cache --ignore-path .gitignore --quiet"
   },
   "engines": {
-    "node": ">=6.0.0"
+    "node": ">=8.0.0"
   },
   "dependencies": {
-    "cosmiconfig": "^5.2.0",
-    "deasync": "^0.1.14",
-    "postcss": "^7.0.14"
+    "cosmiconfig": "^5.2.1",
+    "deasync": "^0.1.15",
+    "postcss": "^7.0.16"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.0",
-    "@babel/preset-env": "^7.4.2",
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
     "babel-eslint": "^10.0.1",
-    "eslint": "^5.15.3",
-    "eslint-config-dev": "^2.0.0",
-    "postcss-tape": "^4.0.0",
+    "eslint": "^5.16.0",
+    "postcss-tape": "^5.0.0",
     "pre-commit": "^1.2.2",
-    "rollup": "^1.7.3",
+    "rollup": "^1.12.3",
     "rollup-plugin-babel": "^4.3.2"
   },
   "eslintConfig": {
-    "extends": "dev",
-    "parser": "babel-eslint"
+    "env": {
+      "browser": true,
+      "es6": true,
+      "node": true
+    },
+    "extends": "eslint:recommended",
+    "parser": "babel-eslint",
+    "parserOptions": {
+      "ecmaVersion": 2018,
+      "impliedStrict": true,
+      "sourceType": "module"
+    },
+    "root": true,
+    "rules": {
+      "no-console": [0, { "allow": ["warn", "error"] }]
+    }
   },
   "keywords": [
     "javascript",
