+++ rollup.config.js
@@ -3,16 +3,17 @@ import babel from 'rollup-plugin-babel';
 export default {
 	input: 'src/index.js',
 	output: [
-		{ file: 'index.js', format: 'cjs', sourcemap: true },
-		{ file: 'index.mjs', format: 'esm', sourcemap: true }
+		{ file: 'index.js', format: 'cjs', sourcemap: true, strict: false },
+		{ file: 'index.mjs', format: 'esm', sourcemap: true, strict: false }
 	],
 	plugins: [
 		babel({
 			presets: [
 				['@babel/env', {
+					corejs: 3,
 					loose: true,
 					modules: false,
-					targets: { node: 6 },
+					targets: { node: 8 },
 					useBuiltIns: 'entry'
 				}]
 			]
