+++ src/lib/requireFromCWD.js
@@ -1,0 +1,4 @@
+// require a module or file in relationship to the current working directory
+export default function requireFromCWD (id) {
+	return require(require.resolve(id, { paths: [process.cwd()] }));
+}
